+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ Changed
 - **General**
     - Upgrade to django-sodar-core v0.10.1-WIP (#1096, #1113, #1118, #1121, #1135)
     - Upgrade to Python v3.8 and Django v3.2 (#1113)
+    - Update project icons (#1113, #1125)
     - Unify ISA-Tab naming (#1082)
     - Upgrade to Chromedriver v89
     - Upgrade to altamISA v0.2.8 (#1099)